requirements.txt
Django==5.0.1
asgiref==3.7.2
pillow==10.3.0
sqlparse==0.4.4
tzdata==2023.4
openai==1.35.9
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
django-cors-headers==4.4.0
python-decouple==3.8